Calendar contentobject fix
git-svn-id: http://svn.projects.ez.no/mcalendar@18 b87a8fb1-9f58-42a2-a0a4-5798444208a8

+++ trunk/classes/calendarCache.php
@@ -98,7 +98,7 @@ class calendarCache {
         }
 
         foreach($toExpire as $expireFile) {
-            eZFileHandler::unlink($expireFile);
+            if (file_exists($expireFile)) eZFileHandler::unlink($expireFile);
             
         }
